requirements.txt
python-barcode
pillow
reportlab
# tkinter